converted to unsigned long long to int

+++ src/libdecomputil.cc
@@ -7,13 +7,14 @@
 
 extern "C" {
 
-unsigned long long seedDecomp2d_polyhedron( float posx, float posy, 
-    unsigned long long nobs, float* obsx, float* obsy,
+int seedDecomp2d_polyhedron( float posx, float posy, 
+    int nobs, float* obsx, float* obsy,
     float local_bbox_x, float local_bbox_y,
     float dilation_radius, 
-    unsigned long long max_npoly,
+    int max_npoly,
     float* poly_px, float* poly_py, float* poly_nx, float* poly_ny
-){
+    )
+{
 
   // Obstacles
   vec_Vec2f obs;
@@ -47,11 +48,11 @@ unsigned long long seedDecomp2d_polyhedron( float posx, float posy,
 
 }
 
-unsigned long long seedDecomp3d_polyhedron( float posx, float posy, float posz,
-    unsigned long long nobs, float* obsx, float* obsy, float* obsz,
+int seedDecomp3d_polyhedron( float posx, float posy, float posz,
+    int nobs, float* obsx, float* obsy, float* obsz,
     float local_bbox_x, float local_bbox_y, float local_bbox_z,
     float dilation_radius, 
-    unsigned long long max_npoly, 
+    int max_npoly, 
     float* poly_px, float* poly_py, float* poly_pz, 
     float* poly_nx, float* poly_ny, float* poly_nz
     )
